fixed db

+++ backend/src/database.py
@@ -15,10 +15,10 @@ class Product(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String(40), nullable=False)
     description = db.Column(db.String(1000), nullable=False)
-    price = db.Column(db.int(10),  nullable=False)
+    price = db.Column(db.Integer(10),  nullable=False)
     street = db.Column(db.String(100), nullable=False)
-    plz = db.Column(db.int(10),  nullable=False)
-    town = db.Column(db.int(100),  nullable=False)
+    plz = db.Column(db.Integer(10),  nullable=False)
+    town = db.Column(db.Integer(100),  nullable=False)
     category = db.Column(db.String(100), nullable=False)
     userID = db.Column(db.Integer, nullable=False)
